move ContributionList to its own file
